+++ InventoryManagement.Application.Contracts/InventoryAgg/InventoryViewModel.cs
@@ -6,7 +6,7 @@ namespace InventoryManagement.Application.Contracts.InventoryAgg
         public long Id{ get; set; }
         public string ProductName { get; set; }
         public long ProductId { get; set; }
-        public long UnitPrice { get; set; }
+        public decimal UnitPrice { get; set; }
         public int Count { get; set; }
         public bool IsInStock { get; set; }
